Struct xcm_builder::test_utils::Assets

pub struct Assets {
    pub fungible: BTreeMap<AssetId, u128, Global>,
    pub non_fungible: BTreeSet<(AssetId, AssetInstance), Global>,
}

List of non-wildcard fungible and non-fungible assets.

Fields
fungible: BTreeMap<AssetId, u128, Global>

The fungible assets.

non_fungible: BTreeSet<(AssetId, AssetInstance), Global>

The non-fungible assets.

Implementations

impl Assets

pub fn new() -> Assets

New value, containing no assets.

pub fn len(&self) -> usize

Total number of distinct assets.

pub fn is_empty(&self) -> bool

Returns true if self contains no assets.

pub fn fungible_assets_iter(&self) -> impl Iterator<Item = MultiAsset>

A borrowing iterator over the fungible assets.

pub fn non_fungible_assets_iter(&self) -> impl Iterator<Item = MultiAsset>

A borrowing iterator over the non-fungible assets.

pub fn into_assets_iter(self) -> impl Iterator<Item = MultiAsset>

A consuming iterator over all assets.

pub fn assets_iter(&self) -> impl Iterator<Item = MultiAsset>

A borrowing iterator over all assets.

pub fn subsume_assets(&mut self, assets: Assets)

Mutate self to contain all given assets, saturating if necessary.

NOTE: Assets are always sorted, allowing us to optimize this function from O(n^2) to O(n).

pub fn subsume(&mut self, asset: MultiAsset)

Mutate self to contain the given asset, saturating if necessary.

Wildcard values of asset do nothing.

pub fn swapped(&mut self, with: Assets) -> Assets

Swaps two mutable Assets, without deinitializing either one.

pub fn prepend_location(&mut self, prepend: &MultiLocation)

Alter any concretely identified assets by prepending the given MultiLocation.

WARNING: For now we consider this infallible and swallow any errors. It is thus the caller’s responsibility to ensure that any internal asset IDs are able to be prepended without overflow.

pub fn reanchor( &mut self, target: &MultiLocation, context: Junctions, maybe_failed_bin: Option<&mut Assets> )

Mutate the assets to be interpreted as the same assets from the perspective of a target chain. The local chain’s context is provided.

Any assets which were unable to be reanchored are introduced into failed_bin.

pub fn contains_asset(&self, asset: &MultiAsset) -> bool

Returns true if asset is contained within self.

pub fn contains_assets(&self, assets: &MultiAssets) -> bool

Returns true if all assets are contained within self.

pub fn contains(&self, assets: &Assets) -> bool

Returns true if all assets are contained within self.

pub fn ensure_contains(&self, assets: &MultiAssets) -> Result<(), TakeError>

Returns an error unless all assets are contained in self. In the case of an error, the first asset in assets which is not wholly in self is returned.

pub fn saturating_take(&mut self, asset: MultiAssetFilter) -> Assets

Mutates self to its original value less mask and returns true iff it contains at least mask.

Returns Ok with the non-wildcard equivalence of mask taken and mutates self to its value minus mask if self contains asset, and return Err otherwise.

pub fn try_take(&mut self, mask: MultiAssetFilter) -> Result<Assets, TakeError>

Mutates self to its original value less mask and returns true iff it contains at least mask.

Returns Ok with the non-wildcard equivalence of asset taken and mutates self to its value minus asset if self contains asset, and return Err otherwise.

pub fn checked_sub(self, asset: MultiAsset) -> Result<Assets, Assets>

Consumes self and returns its original value excluding asset iff it contains at least asset.

pub fn min(&self, mask: &MultiAssetFilter) -> Assets

Return the assets in self, but (asset-wise) of no greater value than mask.

The number of unique assets which are returned will respect the count parameter in the counted wildcard variants of mask.

Example:

use xcm_executor::Assets;
use xcm::latest::prelude::*;
let assets_i_have: Assets = vec![ (Here, 100).into(), ([0; 32], 100).into() ].into();
let assets_they_want: MultiAssetFilter = vec![ (Here, 200).into(), ([0; 32], 50).into() ].into();

let assets_we_can_trade: Assets = assets_i_have.min(&assets_they_want);
assert_eq!(assets_we_can_trade.into_assets_iter().collect::<Vec<_>>(), vec![
	(Here, 100).into(), ([0; 32], 50).into(),
]);