Trait xcm_builder::test_utils::Config

pub trait Config {

    type Call: Parameter + Dispatchable + GetDispatchInfo;
    type XcmSender: SendXcm;
    type AssetTransactor: TransactAsset;
    type OriginConverter: ConvertOrigin<<Self::Call as Dispatchable>::Origin>;
    type IsReserve: FilterAssetLocation;
    type IsTeleporter: FilterAssetLocation;
    type LocationInverter: InvertLocation;
    type Barrier: ShouldExecute;
    type Weigher: WeightBounds<Self::Call>;
    type Trader: WeightTrader;
    type ResponseHandler: OnResponse;
    type AssetTrap: DropAssets;
    type AssetClaims: ClaimAssets;
    type SubscriptionService: VersionChangeNotifier;

}

The trait to parameterize the XcmExecutor.

Associated Types

type Call: Parameter + Dispatchable + GetDispatchInfo

The outer call dispatch type.

type XcmSender: SendXcm

How to send an onward XCM message.

type AssetTransactor: TransactAsset

How to withdraw and deposit an asset.

type OriginConverter: ConvertOrigin<<Self::Call as Dispatchable>::Origin>

How to get a call origin from a OriginKind value.

type IsReserve: FilterAssetLocation

Combinations of (Location, Asset) pairs which we trust as reserves.

type IsTeleporter: FilterAssetLocation

Combinations of (Location, Asset) pairs which we trust as teleporters.

type LocationInverter: InvertLocation

Means of inverting a location.

type Barrier: ShouldExecute

Whether we should execute the given XCM at all.

type Weigher: WeightBounds<Self::Call>

The means of determining an XCM message’s weight.

type Trader: WeightTrader

The means of purchasing weight credit for XCM execution.

type ResponseHandler: OnResponse

What to do when a response of a query is found.

type AssetTrap: DropAssets

The general asset trap - handler for when assets are left in the Holding Register at the end of execution.

type AssetClaims: ClaimAssets

The handler for when there is an instruction to claim assets.

type SubscriptionService: VersionChangeNotifier

How we handle version subscription requests.