Trait xcm_builder::DescribeLocation

pub trait DescribeLocation {
    // Required method
    fn describe_location(location: &MultiLocation) -> Option<Vec<u8>>;
}

Means of converting a location into a stable and unique descriptive identifier.

Required Methods

fn describe_location(location: &MultiLocation) -> Option<Vec<u8>>

Create a description of the given location if possible. No two locations should have the same descriptor.