Crate xcm_builder

XCM-Builder

Types and helpers for building XCM configuration.

Modules

test_utils

Structs

Account32Hash

AccountId32Aliases

Extracts the AccountId32 from the passed location if the network matches.

AccountKey20Aliases

AliasForeignAccountId32

Alias a Foreign AccountId32 with a local AccountId32 if the foreign AccountId32 matches the Prefix pattern.

AliasesIntoAccountId32

Conversion implementation which converts from a [u8; 32]-based AccountId into a MultiLocation consisting solely of a AccountId32 junction with a fixed value for its network (provided by Network) and the AccountId’s [u8; 32] datum for the id.

AllowExplicitUnpaidExecutionFrom

Allows execution from any origin that is contained in T (i.e. T::Contains(origin)) if the message begins with the instruction UnpaidExecution.

AllowKnownQueryResponses

Allows only messages if the generic ResponseHandler expects them via expecting_response.

AllowSubscriptionsFrom

Allows execution from origin if it is just a straight SubscribeVersion or UnsubscribeVersion instruction.

AllowTopLevelPaidExecutionFrom

Allows execution from origin if it is contained in T (i.e. T::Contains(origin)) taking payments into account.

AllowUnpaidExecutionFrom

Allows execution from any origin that is contained in T (i.e. T::Contains(origin)).

AsPrefixedGeneralIndex

Converter struct implementing AssetIdConversion converting a numeric asset ID (must be TryFrom/TryInto<u128>) into a GeneralIndex junction, prefixed by some MultiLocation value. The MultiLocation value will typically be a PalletInstance junction.

BackingToPlurality

Convert implementation to convert from some an origin which implements Backing into a corresponding Plurality MultiLocation.

BridgeBlobDispatcher

BridgeMessage

Case

Accepts an asset if it is contained in the given T’s Get implementation.

ChildParachainAsNative

ChildParachainConvertsVia

ChildSystemParachainAsSuperuser

ConvertedAbstractId

ConvertedConcreteId

CurrencyAdapter

Simple adapter to use a currency as asset transactor. This type can be used as type AssetTransactor in xcm::Config.

DenyReserveTransferToRelayChain

DenyThenTry

Deny executing the XCM if it matches any of the Deny filter regardless of anything else. If it passes the Deny, and matches one of the Allow cases then it is let through.

DescribeAccountId32Terminal

DescribeAccountKey20Terminal

DescribeFamily

DescribePalletTerminal

DescribeTerminus

DualMint

Implementation of AssetChecking which subjects a given set of assets L to having their teleportations recorded with a MintLocation::Local and a second set of assets R to having their teleportations recorded with a MintLocation::NonLocal.

EnsureXcmOrigin

EnsureOrigin barrier to convert from dispatch origin to XCM origin, if one exists.

FixedLocation

Adapter struct which implements a conversion from any AssetKind into a [LocatableAsset] value using a fixed Location for the location field.

FixedRateOfFungible

Simple fee calculator that requires payment in a single fungible at a fixed rate.

FixedWeightBounds

FungiblesAdapter

FungiblesMutateAdapter

FungiblesTransferAdapter

TransactAsset implementation to convert a fungibles implementation to become usable in XCM.

GlobalConsensusConvertsFor

Converts a location which is a top-level relay chain (which provides its own consensus) into a 32-byte AccountId.

GlobalConsensusParachainConvertsFor

Converts a location which is a top-level parachain (i.e. a parachain held on a Relay-chain which provides its own consensus) into a 32-byte AccountId.

HashedDescription

HaulBlobExporter

IsAbstract

Same as IsConcrete but for a fungible with abstract location.

IsChildSystemParachain

Allows a message only if it is from a system-level child parachain.

IsConcrete

Converts a MultiAsset into balance B if it is a concrete fungible with an id equal to that given by T’s Get.

LocalMint

Implementation of AssetChecking which subjects a given set of assets T to having their teleportations recorded with a MintLocation::Local.

LocatableAssetId

Simple struct which contains both an XCM location and asset_id to identify an asset which exists on some chain.

MatchedConvertedConcreteId

Matcher

Struct created from calling fn matcher() on a mutable slice of Instructions.

NativeAsset

Accepts an asset iff it is a native asset.

NetworkExportTable

NoChecking

Implementation of AssetChecking which subjects no assets to having their teleportations recorded.

NonFungiblesAdapter

NonFungiblesMutateAdapter

NonFungiblesTransferAdapter

NonLocalMint

Implementation of AssetChecking which subjects a given set of assets T to having their teleportations recorded with a MintLocation::NonLocal.

OriginToPluralityVoice

Convert implementation to convert from an origin which passes the check of an EnsureOrigin into a voice of a given pluralistic Body.

ParentAsSuperuser

ParentIsPreset

A MultiLocation consisting of a single Parent Junction will be converted to the parent AccountId.

PayOverXcm

Implementation of the frame_support::traits::tokens::Pay trait, to allow for XCM-based payments of a given Balance of some asset ID existing on some chain under ownership of some Interior location of the local chain to a particular Beneficiary. The AssetKind value is not itself bounded (to avoid the issue of needing to wrap some preexisting datatype), however a converter type AssetKindToLocatableAsset must be provided in order to translate it into a LocatableAsset, which comprises both an XCM MultiLocation describing the XCM endpoint on which the asset to be paid resides and an XCM AssetId to identify the specific asset at that endpoint.

ProcessXcmMessage

A message processor that delegates execution to an [XcmExecutor].

RelayChainAsNative

RespectSuspension

Barrier condition that allows for a SuspensionChecker that controls whether or not the XCM executor will be suspended from executing the given XCM.

SiblingParachainAsNative

SiblingParachainConvertsVia

SiblingSystemParachainAsSuperuser

SignedAccountId32AsNative

SignedAccountKey20AsNative

SignedToAccountId32

Convert implementation to convert from some a Signed (system) Origin into an AccountId32.

SovereignPaidRemoteExporter

Implementation of SendXcm which wraps the message inside an ExportMessage instruction and sends it to a destination known to be able to handle it.

SovereignSignedViaLocation

Sovereign accounts use the system’s Signed origin with an account ID derived from the LocationConverter.

TakeWeightCredit

Execution barrier that just takes max_weight from properties.weight_credit.

TrailingSetTopicAsId

Sets the message ID to t using a SetTopic(t) in the last position if present.

UnpaidLocalExporter

Implementation of SendXcm which uses the given ExportXcm implementation in order to forward the message over a bridge.

UnpaidRemoteExporter

Implementation of SendXcm which wraps the message inside an ExportMessage instruction and sends it to a destination known to be able to handle it.

UsingComponents

Weight trader which uses the configured WeightToFee to set the right price for weight and then places any weight bought into the right account.

WeightInfoBounds

WithComputedOrigin

A derivative barrier, which scans the first MaxPrefixes instructions for origin-alterers and then evaluates should_execute of the InnerBarrier based on the remaining instructions and the newly computed origin.

WithTopicSource

Wrapper router which, if the message does not already end with a SetTopic instruction, prepends one to the message filled with an ID from TopicSource. This ID is returned from a successful deliver.

WithUniqueTopic

Wrapper router which, if the message does not already end with a SetTopic instruction, appends one to the message filled with a universally unique ID. This ID is returned from a successful deliver.

Enums

DispatchBlobError

HaulBlobError

MintLocation

The location which is allowed to mint a particular asset.

Traits

AssetChecking

Simple trait to indicate whether an asset is subject to having its teleportation into and out of this chain recorded and if so in what MintLocation.

CreateMatcher

Creates an instruction matcher from an XCM. Since XCM versions differ, we need to make a trait here to unify the interfaces among them.

DescribeLocation

Means of converting a location into a stable and unique descriptive identifier.

DispatchBlob

ExporterFor

HaulBlob

MatchXcm

API that allows to pattern-match against anything that is contained within an XCM.

TakeRevenue

Function trait for handling some revenue. Similar to a negative imbalance (credit) handler, but for a MultiAsset. Sensible implementations will deposit the asset in some known treasury or block-author account.

Functions

ensure_is_remote

Returns the network ID and consensus location within that network of the remote location dest which is itself specified as a location relative to the local chain, itself situated at universal_local within the consensus universe. If dest is not a location in remote consensus, then an error is returned.

Type Definitions

ConvertedAbstractAssetIdDeprecated

ConvertedConcreteAssetIdDeprecated

DescribeAccountIdTerminal

DescribeAllTerminal

ForeignChainAliasAccountDeprecated

This is deprecated in favor of the more modular HashedDescription converter. If your chain has previously used this, then you can retain backwards compatibility using HashedDescription and a tuple with LegacyDescribeForeignChainAccount as the first element. For example:

PayAccountId32OnChainOverXcm

Specialization of the PayOverXcm trait to allow [u8; 32]-based AccountId values to be paid on a remote chain.